+++ src/Commands/GenerateDocumentation.php
@@ -6,6 +6,8 @@ use ReflectionClass;
 use ReflectionException;
 use Illuminate\Routing\Route;
 use Illuminate\Console\Command;
+use Mpociot\ApiDoc\Tools\Flags;
+use Mpociot\ApiDoc\Tools\Utils;
 use Mpociot\Reflection\DocBlock;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\URL;
@@ -13,6 +15,7 @@ use Mpociot\ApiDoc\Tools\Generator;
 use Mpociot\ApiDoc\Tools\RouteMatcher;
 use Mpociot\Documentarian\Documentarian;
 use Mpociot\ApiDoc\Postman\CollectionWriter;
+use Mpociot\ApiDoc\Tools\DocumentationConfig;
 
 class GenerateDocumentation extends Command
 {
@@ -34,6 +37,11 @@ class GenerateDocumentation extends Command
 
     private $routeMatcher;
 
+    /**
+     * @var DocumentationConfig
+     */
+    private $docConfig;
+
     public function __construct(RouteMatcher $routeMatcher)
     {
         parent::__construct();
@@ -47,15 +55,26 @@ class GenerateDocumentation extends Command
      */
     public function handle()
     {
-        URL::forceRootUrl(config('app.url'));
-        $usingDingoRouter = strtolower(config('apidoc.router')) == 'dingo';
-        if ($usingDingoRouter) {
-            $routes = $this->routeMatcher->getDingoRoutesToBeDocumented(config('apidoc.routes'));
-        } else {
-            $routes = $this->routeMatcher->getLaravelRoutesToBeDocumented(config('apidoc.routes'));
+        // Using a global static variable here, so fuck off if you don't like it
+        // Also, the --verbose option is included with all Artisan commands
+        Flags::$shouldBeVerbose = $this->option('verbose');
+
+        $this->docConfig = new DocumentationConfig(config('apidoc'));
+
+        try {
+            URL::forceRootUrl($this->docConfig->get('base_url'));
+        } catch (\Error $e) {
+            echo "Warning: Couldn't force base url as your version of Lumen doesn't have the forceRootUrl method.\n";
+            echo "You should probably double check URLs in your generated documentation.\n";
         }
 
-        $generator = new Generator(config('apidoc.faker_seed'));
+        $usingDingoRouter = strtolower($this->docConfig->get('router')) == 'dingo';
+        $routes = $this->docConfig->get('routes');
+        $routes = $usingDingoRouter
+            ? $this->routeMatcher->getDingoRoutesToBeDocumented($routes)
+            : $this->routeMatcher->getLaravelRoutesToBeDocumented($routes);
+
+        $generator = new Generator($this->docConfig);
         $parsedRoutes = $this->processRoutes($generator, $routes);
         $parsedRoutes = collect($parsedRoutes)->groupBy('group')
             ->sortBy(static function ($group) {
@@ -73,7 +92,7 @@ class GenerateDocumentation extends Command
      */
     private function writeMarkdown($parsedRoutes)
     {
-        $outputPath = config('apidoc.output');
+        $outputPath = $this->docConfig->get('output');
         $targetFile = $outputPath.DIRECTORY_SEPARATOR.'source'.DIRECTORY_SEPARATOR.'index.md';
         $compareFile = $outputPath.DIRECTORY_SEPARATOR.'source'.DIRECTORY_SEPARATOR.'.compare.md';
         $prependFile = $outputPath.DIRECTORY_SEPARATOR.'source'.DIRECTORY_SEPARATOR.'prepend.md';
@@ -83,7 +102,7 @@ class GenerateDocumentation extends Command
             ->with('outputPath', ltrim($outputPath, 'public/'))
             ->with('showPostmanCollectionButton', $this->shouldGeneratePostmanCollection());
 
-        $settings = ['languages' => config('apidoc.example_languages')];
+        $settings = ['languages' => $this->docConfig->get('example_languages')];
         $parsedRouteOutput = $parsedRoutes->map(function ($routeGroup) use ($settings) {
             return $routeGroup->map(function ($route) use ($settings) {
                 if (count($route['cleanBodyParameters']) && ! isset($route['headers']['Content-Type'])) {
@@ -92,6 +111,7 @@ class GenerateDocumentation extends Command
                 $route['output'] = (string) view('apidoc::partials.route')
                     ->with('route', $route)
                     ->with('settings', $settings)
+                    ->with('baseUrl', $this->docConfig->get('base_url'))
                     ->render();
 
                 return $route;
@@ -144,7 +164,7 @@ class GenerateDocumentation extends Command
             ->with('infoText', $infoText)
             ->with('prependMd', $prependFileContents)
             ->with('appendMd', $appendFileContents)
-            ->with('outputPath', config('apidoc.output'))
+            ->with('outputPath', $this->docConfig->get('output'))
             ->with('showPostmanCollectionButton', $this->shouldGeneratePostmanCollection())
             ->with('parsedRoutes', $parsedRouteOutput);
 
@@ -162,7 +182,7 @@ class GenerateDocumentation extends Command
             ->with('infoText', $infoText)
             ->with('prependMd', $prependFileContents)
             ->with('appendMd', $appendFileContents)
-            ->with('outputPath', config('apidoc.output'))
+            ->with('outputPath', $this->docConfig->get('output'))
             ->with('showPostmanCollectionButton', $this->shouldGeneratePostmanCollection())
             ->with('parsedRoutes', $parsedRouteOutput);
 
@@ -182,7 +202,7 @@ class GenerateDocumentation extends Command
             file_put_contents($outputPath.DIRECTORY_SEPARATOR.'collection.json', $this->generatePostmanCollection($parsedRoutes));
         }
 
-        if ($logo = config('apidoc.logo')) {
+        if ($logo = $this->docConfig->get('logo')) {
             copy(
                 $logo,
                 $outputPath.DIRECTORY_SEPARATOR.'images'.DIRECTORY_SEPARATOR.'logo.png'
@@ -202,7 +222,7 @@ class GenerateDocumentation extends Command
         foreach ($routes as $routeItem) {
             $route = $routeItem['route'];
             /** @var Route $route */
-            if ($this->isValidRoute($route) && $this->isRouteVisibleForDocumentation($route->getAction()['uses'])) {
+            if ($this->isValidRoute($route) && $this->isRouteVisibleForDocumentation($route->getAction())) {
                 $parsedRoutes[] = $generator->processRoute($route, $routeItem['apply']);
                 $this->info('Processed route: ['.implode(',', $generator->getMethods($route)).'] '.$generator->getUri($route));
             } else {
@@ -220,19 +240,24 @@ class GenerateDocumentation extends Command
      */
     private function isValidRoute(Route $route)
     {
-        return ! is_callable($route->getAction()['uses']) && ! is_null($route->getAction()['uses']);
+        $action = Utils::getRouteActionUses($route->getAction());
+        if (is_array($action)) {
+            $action = implode('@', $action);
+        }
+
+        return ! is_callable($action) && ! is_null($action);
     }
 
     /**
-     * @param $route
+     * @param $action
      *
      * @throws ReflectionException
      *
      * @return bool
      */
-    private function isRouteVisibleForDocumentation($route)
+    private function isRouteVisibleForDocumentation($action)
     {
-        list($class, $method) = explode('@', $route);
+        list($class, $method) = Utils::getRouteActionUses($action);
         $reflection = new ReflectionClass($class);
 
         if (! $reflection->hasMethod($method)) {
@@ -245,7 +270,7 @@ class GenerateDocumentation extends Command
             $phpdoc = new DocBlock($comment);
 
             return collect($phpdoc->getTags())
-                ->filter(function ($tag) use ($route) {
+                ->filter(function ($tag) use ($action) {
                     return $tag->getName() === 'hideFromAPIDocumentation';
                 })
                 ->isEmpty();
@@ -263,7 +288,7 @@ class GenerateDocumentation extends Command
      */
     private function generatePostmanCollection(Collection $routes)
     {
-        $writer = new CollectionWriter($routes);
+        $writer = new CollectionWriter($routes, $this->docConfig->get('base_url'));
 
         return $writer->getCollection();
     }
@@ -275,6 +300,6 @@ class GenerateDocumentation extends Command
      */
     private function shouldGeneratePostmanCollection()
     {
-        return config('apidoc.postman.enabled', is_bool(config('apidoc.postman')) ? config('apidoc.postman') : false);
+        return $this->docConfig->get('postman.enabled', is_bool($this->docConfig->get('postman')) ? $this->docConfig->get('postman') : false);
     }
 }