+++ src/Postman/CollectionWriter.php
@@ -14,18 +14,29 @@ class CollectionWriter
     private $routeGroups;
 
     /**
+     * @var string
+     */
+    private $baseUrl;
+
+    /**
      * CollectionWriter constructor.
      *
      * @param Collection $routeGroups
      */
-    public function __construct(Collection $routeGroups)
+    public function __construct(Collection $routeGroups, $baseUrl)
     {
         $this->routeGroups = $routeGroups;
+        $this->baseUrl = $baseUrl;
     }
 
     public function getCollection()
     {
-        URL::forceRootUrl(config('app.url'));
+        try {
+            URL::forceRootUrl($this->baseUrl);
+        } catch (\Error $e) {
+            echo "Warning: Couldn't force base url as your version of Lumen doesn't have the forceRootUrl method.\n";
+            echo "You should probably double check URLs in your generated Postman collection.\n";
+        }
 
         $collection = [
             'variables' => [],
@@ -45,14 +56,29 @@ class CollectionWriter
                         return [
                             'name' => $route['title'] != '' ? $route['title'] : url($route['uri']),
                             'request' => [
-                                'url' => url($route['uri']),
+                                'url' => url($route['uri']).(collect($route['queryParameters'])->isEmpty()
+                                    ? ''
+                                    : ('?'.implode('&', collect($route['queryParameters'])->map(function ($parameter, $key) {
+                                        return $key.'='.($parameter['value'] ?? '');
+                                    })->all()))),
                                 'method' => $route['methods'][0],
+                                'header' => collect($route['headers'])
+                                    ->union([
+                                        'Accept' => 'application/json',
+                                    ])
+                                    ->map(function ($value, $header) {
+                                        return [
+                                            'key' => $header,
+                                            'value' => $value,
+                                        ];
+                                    })
+                                    ->values()->all(),
                                 'body' => [
                                     'mode' => $mode,
                                     $mode => collect($route['bodyParameters'])->map(function ($parameter, $key) {
                                         return [
                                             'key' => $key,
-                                            'value' => isset($parameter['value']) ? $parameter['value'] : '',
+                                            'value' => $parameter['value'] ?? '',
                                             'type' => 'text',
                                             'enabled' => true,
                                         ];