+++ src/Tools/Utils.php
@@ -4,6 +4,8 @@ namespace Mpociot\ApiDoc\Tools;
 
 use Illuminate\Support\Str;
 use Illuminate\Routing\Route;
+use League\Flysystem\Filesystem;
+use League\Flysystem\Adapter\Local;
 
 class Utils
 {
@@ -15,6 +17,28 @@ class Utils
     }
 
     /**
+     * @param array $action
+     *
+     * @return array|null
+     */
+    public static function getRouteActionUses(array $action)
+    {
+        if ($action['uses'] !== null) {
+            if (is_array($action['uses'])) {
+                return $action['uses'];
+            } elseif (is_string($action['uses'])) {
+                return explode('@', $action['uses']);
+            }
+        }
+        if (array_key_exists(0, $action) && array_key_exists(1, $action)) {
+            return [
+                0 => $action[0],
+                1 => $action[1],
+            ];
+        }
+    }
+
+    /**
      * Transform parameters in URLs into real values (/users/{user} -> /users/2).
      * Uses bindings specified by caller, otherwise just uses '1'.
      *
@@ -23,7 +47,7 @@ class Utils
      *
      * @return mixed
      */
-    protected static function replaceUrlParameterBindings(string $uri, array $bindings)
+    public static function replaceUrlParameterBindings(string $uri, array $bindings)
     {
         foreach ($bindings as $path => $binding) {
             // So we can support partial bindings like
@@ -40,5 +64,25 @@ class Utils
         $uri = preg_replace('/{(.+?)}/', 1, $uri);
 
         return $uri;
+    }
+
+    public function dumpException(\Exception $e)
+    {
+        if (class_exists(\NunoMaduro\Collision\Handler::class)) {
+            $handler = new \NunoMaduro\Collision\Handler;
+            $handler->setInspector(new \Whoops\Exception\Inspector($e));
+            $handler->setException($e);
+            $handler->handle();
+        } else {
+            dump($e);
+            echo "You can get better exception output by installing the library \nunomaduro/collision (PHP 7.1+ only).\n";
+        }
+    }
+
+    public static function deleteDirectoryAndContents($dir)
+    {
+        $adapter = new Local(realpath(__DIR__.'/../../'));
+        $fs = new Filesystem($adapter);
+        $fs->deleteDir($dir);
     }
 }